fix: close remaining pr1300 follow-ups
