+++ packages/web/src/components/SessionDetail.tsx
@@ -1,32 +1,38 @@
 "use client";
 
-import { useState, useEffect, useRef, useMemo, useCallback, type ReactNode } from "react";
+import { useCallback, useEffect, useMemo, useState } from "react";
+import dynamic from "next/dynamic";
 import { useSearchParams } from "next/navigation";
+import { cn } from "@/lib/cn";
+import { getSessionTitle } from "@/lib/format";
 import { useMediaQuery, MOBILE_BREAKPOINT } from "@/hooks/useMediaQuery";
+import type { ProjectInfo } from "@/lib/project-name";
 import {
   type DashboardSession,
-  type DashboardPR,
   NON_RESTORABLE_STATUSES,
   getActivitySignalLabel,
-  isPRMergeReady,
-  isPRRateLimited,
-  isPRUnenriched,
   isDashboardRuntimeEnded,
   isDashboardSessionRestorable,
 } from "@/lib/types";
-import { CI_STATUS } from "@aoagents/ao-core/types";
-import { cn } from "@/lib/cn";
-import dynamic from "next/dynamic";
-import { getSessionTitle } from "@/lib/format";
-import type { ProjectInfo } from "@/lib/project-name";
-
 import { MobileBottomNav } from "./MobileBottomNav";
 import { ProjectSidebar } from "./ProjectSidebar";
-import { SessionTruthPanel } from "./SessionTruthPanel";
+import { SessionDetailPRCard } from "./SessionDetailPRCard";
+import { SessionTopStrip } from "./SessionDetailTopStrip";
 import { SessionReportAuditPanel } from "./SessionReportAuditPanel";
+import { SessionTruthPanel } from "./SessionTruthPanel";
+import { askAgentToFix } from "./session-detail-agent-actions";
+import {
+  activityToneClass,
+  ciToneClass,
+  formatTimeCompact,
+  getCiShortLabel,
+  getReviewShortLabel,
+  mobileStatusPillClass,
+  sessionActivityMeta,
+} from "./session-detail-utils";
 
 const DirectTerminal = dynamic(
-  () => import("./DirectTerminal").then((m) => ({ default: m.DirectTerminal })),
+  () => import("./DirectTerminal").then((module) => ({ default: module.DirectTerminal })),
   {
     ssr: false,
     loading: () => (
@@ -56,418 +62,6 @@ interface SessionDetailProps {
   onRetrySidebar?: () => void;
 }
 
-// ── Helpers ──────────────────────────────────────────────────────────
-
-function formatTimeCompact(isoDate: string | null): string {
-  if (!isoDate) return "just now";
-  const ts = new Date(isoDate).getTime();
-  if (!Number.isFinite(ts)) return "just now";
-  const diffMs = Date.now() - ts;
-  if (diffMs <= 0) return "just now";
-  const diffMins = Math.floor(diffMs / 60000);
-  const diffHours = Math.floor(diffMins / 60);
-  if (diffMins < 1) return "just now";
-  if (diffMins < 60) return `${diffMins}m ago`;
-  if (diffHours < 24) return `${diffHours}h ago`;
-  return `${Math.floor(diffHours / 24)}d ago`;
-}
-
-function getCiShortLabel(pr: DashboardPR): string {
-  if (isPRRateLimited(pr) || isPRUnenriched(pr)) return "CI";
-  if (pr.ciStatus === "passing") return "CI passing";
-  if (pr.ciStatus === "failing") return "CI failed";
-  return "CI pending";
-}
-
-function getReviewShortLabel(pr: DashboardPR): string {
-  if (isPRRateLimited(pr) || isPRUnenriched(pr)) return "";
-  if (pr.reviewDecision === "approved") return "approved";
-  if (pr.reviewDecision === "changes_requested") return "changes";
-  return "review";
-}
-
-const activityMeta: Record<string, { label: string; color: string }> = {
-  active: { label: "Active", color: "var(--color-status-working)" },
-  ready: { label: "Ready", color: "var(--color-status-ready)" },
-  idle: { label: "Idle", color: "var(--color-status-idle)" },
-  waiting_input: { label: "Waiting for input", color: "var(--color-status-attention)" },
-  blocked: { label: "Blocked", color: "var(--color-status-error)" },
-  exited: { label: "Exited", color: "var(--color-status-error)" },
-};
-
-function cleanBugbotComment(body: string): { title: string; description: string } {
-  const isBugbot = body.includes("<!-- DESCRIPTION START -->") || body.includes("### ");
-  if (isBugbot) {
-    const titleMatch = body.match(/###\s+(.+?)(?:\n|$)/);
-    const title = titleMatch ? titleMatch[1].replace(/\*\*/g, "").trim() : "Comment";
-    const descMatch = body.match(
-      /<!-- DESCRIPTION START -->\s*([\s\S]*?)\s*<!-- DESCRIPTION END -->/,
-    );
-    const description = descMatch ? descMatch[1].trim() : body.split("\n")[0] || "No description";
-    return { title, description };
-  }
-  return { title: "Comment", description: body.trim() };
-}
-
-function buildGitHubBranchUrl(pr: DashboardPR): string {
-  return `https://github.com/${pr.owner}/${pr.repo}/tree/${pr.branch}`;
-}
-
-function activityStateClass(activityLabel: string): string {
-  const normalized = activityLabel.toLowerCase();
-  if (normalized === "active") return "session-detail-status-pill--active";
-  if (normalized === "ready") return "session-detail-status-pill--ready";
-  if (normalized === "idle") return "session-detail-status-pill--idle";
-  if (normalized === "waiting for input") return "session-detail-status-pill--waiting";
-  if (normalized === "blocked" || normalized === "exited") {
-    return "session-detail-status-pill--error";
-  }
-  return "session-detail-status-pill--neutral";
-}
-
-function activityToneClass(activityColor: string): string {
-  switch (activityColor) {
-    case "var(--color-status-working)":
-      return "session-detail-tone--working";
-    case "var(--color-status-ready)":
-      return "session-detail-tone--ready";
-    case "var(--color-status-idle)":
-      return "session-detail-tone--idle";
-    case "var(--color-status-attention)":
-      return "session-detail-tone--attention";
-    case "var(--color-status-error)":
-      return "session-detail-tone--error";
-    default:
-      return "session-detail-tone--muted";
-  }
-}
-
-function mobileStatusPillClass(activityLabel: string): string {
-  const normalized = activityLabel.toLowerCase();
-  if (normalized === "active") return "session-detail__status-pill--active";
-  if (normalized === "ready") return "session-detail__status-pill--ready";
-  if (normalized === "waiting for input") return "session-detail__status-pill--waiting";
-  if (normalized === "blocked" || normalized === "exited") {
-    return "session-detail__status-pill--error";
-  }
-  return "session-detail__status-pill--idle";
-}
-
-function ciToneClass(pr: DashboardPR): string {
-  if (isPRRateLimited(pr) || isPRUnenriched(pr)) return "session-detail-ci-tone--neutral";
-  if (pr.ciStatus === "passing") return "session-detail-ci-tone--pass";
-  if (pr.ciStatus === "failing") return "session-detail-ci-tone--fail";
-  return "session-detail-ci-tone--pending";
-}
-
-function SessionTopStrip({
-  headline,
-  crumbId,
-  activityLabel,
-  activityColor,
-  branch,
-  pr,
-  isOrchestrator = false,
-  crumbHref,
-  crumbLabel,
-  rightSlot,
-  onKill,
-  onRestore,
-}: {
-  headline: string;
-  crumbId: string;
-  activityLabel: string;
-  activityColor: string;
-  branch: string | null;
-  pr: DashboardPR | null;
-  isOrchestrator?: boolean;
-  crumbHref: string;
-  crumbLabel: string;
-  rightSlot?: ReactNode;
-  onKill?: () => void;
-  onRestore?: () => void;
-}) {
-  return (
-    <div className="session-detail-top-strip">
-      {/* Breadcrumbs */}
-      <div className="session-detail-crumbs">
-        <a
-          href={crumbHref}
-          className="session-detail-crumb-back"
-        >
-          <svg
-            className="h-3 w-3 opacity-60"
-            fill="none"
-            stroke="currentColor"
-            strokeWidth="2.5"
-            viewBox="0 0 24 24"
-          >
-            <path d="M15 18l-6-6 6-6" />
-          </svg>
-          {crumbLabel}
-        </a>
-        <span className="session-detail-crumb-sep">/</span>
-        <span className="session-detail-crumb-id">{crumbId}</span>
-        {isOrchestrator ? (
-          <span className="session-detail-mode-badge">orchestrator</span>
-        ) : null}
-      </div>
-
-      {/* Identity strip */}
-      <div className="session-detail-identity">
-        <div className="session-detail-identity__info">
-          <h1 className="session-detail-identity__title">
-            {headline}
-          </h1>
-          <div className="session-detail-identity__pills">
-            <div
-              className={cn(
-                "session-detail-status-pill",
-                activityStateClass(activityLabel),
-              )}
-            >
-              <span
-                className={cn(
-                  "session-detail-status-pill__dot",
-                  activityToneClass(activityColor),
-                )}
-              />
-              <span className="session-detail-status-pill__label">
-                {activityLabel}
-              </span>
-            </div>
-            {branch ? (
-              pr ? (
-                <a
-                  href={buildGitHubBranchUrl(pr)}
-                  target="_blank"
-                  rel="noopener noreferrer"
-                  className="session-detail-link-pill session-detail-link-pill--branch session-detail-link-pill--branch-link hover:no-underline"
-                >
-                  {branch}
-                </a>
-              ) : (
-                <span className="session-detail-link-pill session-detail-link-pill--branch">
-                  {branch}
-                </span>
-              )
-            ) : null}
-            {pr ? (
-              <a
-                href={pr.url}
-                target="_blank"
-                rel="noopener noreferrer"
-                className="session-detail-link-pill session-detail-link-pill--pr hover:no-underline"
-              >
-                PR #{pr.number}
-              </a>
-            ) : null}
-            {pr && (pr.additions > 0 || pr.deletions > 0) ? (
-              <span className="session-detail-link-pill session-detail-link-pill--diff">
-                <span className="session-detail-diff--add">+{pr.additions}</span>
-                {" "}
-                <span className="session-detail-diff--del">-{pr.deletions}</span>
-              </span>
-            ) : null}
-          </div>
-        </div>
-
-        {rightSlot ? (
-          <div className="session-detail-identity__actions session-detail-identity__actions--custom">
-            {rightSlot}
-          </div>
-        ) : (
-          <div className="session-detail-identity__actions">
-            {onRestore ? (
-              <button
-                type="button"
-                className="done-restore-btn"
-                onClick={onRestore}
-              >
-                <svg
-                  fill="none"
-                  stroke="currentColor"
-                  strokeWidth="2"
-                  viewBox="0 0 24 24"
-                  className="h-3 w-3"
-                >
-                  <polyline points="1 4 1 10 7 10" />
-                  <path d="M3.51 15a9 9 0 1 0 2.13-9.36L1 10" />
-                </svg>
-                Restore
-              </button>
-            ) : onKill ? (
-              <button
-                type="button"
-                className="session-detail-action-btn session-detail-action-btn--danger"
-                onClick={onKill}
-              >
-                <svg className="h-3.5 w-3.5" fill="none" stroke="currentColor" strokeWidth="2" viewBox="0 0 24 24">
-                  <path d="M3 6h18M19 6v14a2 2 0 0 1-2 2H7a2 2 0 0 1-2-2V6m3 0V4a2 2 0 0 1 2-2h4a2 2 0 0 1 2 2v2" />
-                </svg>
-                Kill
-              </button>
-            ) : null}
-          </div>
-        )}
-      </div>
-    </div>
-  );
-}
-
-async function askAgentToFix(
-  sessionId: string,
-  comment: { url: string; path: string; body: string },
-  onSuccess: () => void,
-  onError: () => void,
-) {
-  try {
-    const { title, description } = cleanBugbotComment(comment.body);
-    const message = `Please address this review comment:\n\nFile: ${comment.path}\nComment: ${title}\nDescription: ${description}\n\nComment URL: ${comment.url}\n\nAfter fixing, mark the comment as resolved at ${comment.url}`;
-    const res = await fetch(`/api/sessions/${encodeURIComponent(sessionId)}/message`, {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({ message }),
-    });
-    if (!res.ok) throw new Error(`HTTP ${res.status}`);
-    onSuccess();
-  } catch (err) {
-    console.error("Failed to send message to agent:", err);
-    onError();
-  }
-}
-
-// ── Orchestrator status strip ─────────────────────────────────────────
-
-function _OrchestratorStatusStrip({
-  zones,
-  createdAt,
-  headline,
-  activityLabel,
-  activityColor,
-  branch,
-  pr,
-  crumbHref,
-  crumbLabel,
-}: {
-  zones: OrchestratorZones;
-  createdAt: string;
-  headline: string;
-  activityLabel: string;
-  activityColor: string;
-  branch: string | null;
-  pr: DashboardPR | null;
-  crumbHref: string;
-  crumbLabel: string;
-}) {
-  const [uptime, setUptime] = useState<string>("");
-
-  useEffect(() => {
-    const compute = () => {
-      const diff = Date.now() - new Date(createdAt).getTime();
-      const h = Math.floor(diff / 3_600_000);
-      const m = Math.floor((diff % 3_600_000) / 60_000);
-      setUptime(h > 0 ? `${h}h ${m}m` : `${m}m`);
-    };
-    compute();
-    const id = setInterval(compute, 30_000);
-    return () => clearInterval(id);
-  }, [createdAt]);
-
-  const stats: Array<{ value: number; label: string; toneClass: string }> = [
-    {
-      value: zones.merge,
-      label: "merge-ready",
-      toneClass: "session-detail-zone-pill--merge",
-    },
-    {
-      value: zones.respond,
-      label: "responding",
-      toneClass: "session-detail-zone-pill--respond",
-    },
-    {
-      value: zones.review,
-      label: "review",
-      toneClass: "session-detail-zone-pill--review",
-    },
-    {
-      value: zones.working,
-      label: "working",
-      toneClass: "session-detail-zone-pill--working",
-    },
-    {
-      value: zones.pending,
-      label: "pending",
-      toneClass: "session-detail-zone-pill--pending",
-    },
-    {
-      value: zones.done,
-      label: "done",
-      toneClass: "session-detail-zone-pill--done",
-    },
-  ].filter((s) => s.value > 0);
-
-  const total =
-    zones.merge + zones.respond + zones.review + zones.working + zones.pending + zones.done;
-
-  return (
-    <div className="mx-auto max-w-[1180px] px-5 pt-5 lg:px-8">
-      <SessionTopStrip
-        headline={headline}
-        crumbId={headline}
-        activityLabel={activityLabel}
-        activityColor={activityColor}
-        branch={branch}
-        pr={pr}
-        isOrchestrator
-        crumbHref={crumbHref}
-        crumbLabel={crumbLabel}
-        rightSlot={
-          <div className="flex flex-wrap items-center gap-3 lg:justify-end">
-            <div className="flex items-baseline gap-1.5 mr-2">
-              <span className="text-[22px] font-bold leading-none tabular-nums text-[var(--color-text-primary)]">
-                {total}
-              </span>
-              <span className="text-[11px] text-[var(--color-text-tertiary)]">agents</span>
-            </div>
-
-            <div className="h-5 w-px bg-[var(--color-border-subtle)] mr-1" />
-
-            {/* Per-zone pills */}
-            {stats.length > 0 ? (
-              stats.map((s) => (
-                <div
-                  key={s.label}
-                  className={cn("session-detail-zone-pill", s.toneClass)}
-                >
-                  <span className="session-detail-zone-pill__value">
-                    {s.value}
-                  </span>
-                  <span className="session-detail-zone-pill__label">
-                    {s.label}
-                  </span>
-                </div>
-              ))
-            ) : (
-              <span className="text-[12px] text-[var(--color-text-tertiary)]">
-                no active agents
-              </span>
-            )}
-
-            {uptime && (
-              <span className="ml-auto font-[var(--font-mono)] text-[11px] text-[var(--color-text-tertiary)]">
-                up {uptime}
-              </span>
-            )}
-          </div>
-        }
-      />
-    </div>
-  );
-}
-
-// ── Main component ────────────────────────────────────────────────────
-
 export function SessionDetail({
   session,
   isOrchestrator = false,
@@ -490,14 +84,12 @@ export function SessionDetail({
     !isOrchestrator &&
     isDashboardSessionRestorable(session) &&
     !NON_RESTORABLE_STATUSES.has(session.status);
-  const activity = (session.activity && activityMeta[session.activity]) ?? {
+  const activity = (session.activity && sessionActivityMeta[session.activity]) ?? {
     label: getActivitySignalLabel(session),
     color: "var(--color-text-muted)",
   };
   const headline = getSessionTitle(session);
-
   const terminalVariant = isOrchestrator ? "orchestrator" : "agent";
-
   const terminalHeight = isOrchestrator
     ? "clamp(400px, 52vh, 620px)"
     : "clamp(380px, 48vh, 560px)";
@@ -515,37 +107,38 @@ export function SessionDetail({
     : undefined;
   const dashboardHref = session.projectId ? `/?project=${encodeURIComponent(session.projectId)}` : "/";
   const prsHref = session.projectId ? `/prs?project=${encodeURIComponent(session.projectId)}` : "/prs";
-  const crumbHref = dashboardHref;
-  const crumbLabel = "Dashboard";
-
-  const handleKill = useCallback(async () => {
-    try {
-      const res = await fetch(`/api/sessions/${encodeURIComponent(session.id)}/kill`, { method: "POST" });
-      if (!res.ok) throw new Error(`HTTP ${res.status}`);
-      window.location.reload();
-    } catch (err) {
-      console.error("Failed to kill session:", err);
-    }
-  }, [session.id]);
-
-  const handleRestore = useCallback(async () => {
-    try {
-      const res = await fetch(`/api/sessions/${encodeURIComponent(session.id)}/restore`, { method: "POST" });
-      if (!res.ok) throw new Error(`HTTP ${res.status}`);
-      window.location.reload();
-    } catch (err) {
-      console.error("Failed to restore session:", err);
-    }
-  }, [session.id]);
   const headerProjectLabel =
     projects.find((project) => project.id === session.projectId)?.name ?? session.projectId;
-  const showHeaderProjectLabel =
-    headerProjectLabel.trim().toLowerCase() !== "agent orchestrator";
+  const showHeaderProjectLabel = headerProjectLabel.trim().toLowerCase() !== "agent orchestrator";
   const orchestratorHref = useMemo(() => {
     if (isOrchestrator) return `/sessions/${encodeURIComponent(session.id)}`;
     if (!projectOrchestratorId) return null;
     return `/sessions/${encodeURIComponent(projectOrchestratorId)}`;
   }, [isOrchestrator, projectOrchestratorId, session.id]);
+
+  const handleKill = useCallback(async () => {
+    try {
+      const response = await fetch(`/api/sessions/${encodeURIComponent(session.id)}/kill`, {
+        method: "POST",
+      });
+      if (!response.ok) throw new Error(`HTTP ${response.status}`);
+      window.location.reload();
+    } catch (error) {
+      console.error("Failed to kill session:", error);
+    }
+  }, [session.id]);
+
+  const handleRestore = useCallback(async () => {
+    try {
+      const response = await fetch(`/api/sessions/${encodeURIComponent(session.id)}/restore`, {
+        method: "POST",
+      });
+      if (!response.ok) throw new Error(`HTTP ${response.status}`);
+      window.location.reload();
+    } catch (error) {
+      console.error("Failed to restore session:", error);
+    }
+  }, [session.id]);
 
   useEffect(() => {
     const frame = window.requestAnimationFrame(() => setShowTerminal(true));
@@ -639,7 +232,7 @@ export function SessionDetail({
             <main className="session-detail-page min-h-0 flex-1 overflow-y-auto bg-[var(--color-bg-base)]">
               <div className="session-detail-layout">
                 <main className="min-w-0">
-                  {(!isOrchestrator || (isOrchestrator && orchestratorZones)) && (
+                  {(!isOrchestrator || orchestratorZones) && (
                     <SessionTopStrip
                       headline={headline}
                       crumbId={session.id}
@@ -648,8 +241,8 @@ export function SessionDetail({
                       branch={isOrchestrator ? null : session.branch}
                       pr={pr}
                       isOrchestrator={isOrchestrator}
-                      crumbHref={crumbHref}
-                      crumbLabel={crumbLabel}
+                      crumbHref={dashboardHref}
+                      crumbLabel="Dashboard"
                       onKill={isOrchestrator || terminalEnded ? undefined : handleKill}
                       onRestore={isOrchestrator || !isRestorable ? undefined : handleRestore}
                     />
@@ -659,9 +252,11 @@ export function SessionDetail({
                     <section id="session-pr-section" className="session-detail-pr-section">
                       <SessionDetailPRCard
                         pr={pr}
-                        sessionId={session.id}
                         metadata={session.metadata}
                         lifecyclePrReason={session.lifecycle?.prReason}
+                        onAskAgentToFix={(comment, onSuccess, onError) =>
+                          askAgentToFix(session.id, comment, onSuccess, onError)
+                        }
                       />
                     </section>
                   ) : null}
@@ -685,9 +280,7 @@ export function SessionDetail({
                             : activityToneClass(activity.color),
                         )}
                       />
-                      <span className="session-detail-section-label__text">
-                        Live Terminal
-                      </span>
+                      <span className="session-detail-section-label__text">Live Terminal</span>
                     </div>
                     {!showTerminal ? (
                       <div
@@ -725,37 +318,33 @@ export function SessionDetail({
 
   return (
     <div className="session-detail--terminal-first">
-      {/* Floating header */}
       <div className="session-detail__floating-header">
-        <a href={crumbHref} className="session-detail__back" aria-label="Back to dashboard">
-          <svg width="18" height="18" fill="none" stroke="currentColor" strokeWidth="2" viewBox="0 0 24 24">
+        <a href={dashboardHref} className="session-detail__back" aria-label="Back to dashboard">
+          <svg
+            width="18"
+            height="18"
+            fill="none"
+            stroke="currentColor"
+            strokeWidth="2"
+            viewBox="0 0 24 24"
+          >
             <path d="M19 12H5M12 19l-7-7 7-7" />
           </svg>
         </a>
         <span className={cn("session-detail__status-dot", activityToneClass(activity.color))} />
         <span className="session-detail__session-id">{session.id}</span>
-        <span
-          className={cn(
-            "session-detail__status-pill",
-            mobileStatusPillClass(activity.label),
-          )}
-        >
+        <span className={cn("session-detail__status-pill", mobileStatusPillClass(activity.label))}>
           {activity.label.toLowerCase()}
         </span>
-        <span className="session-detail__time">
-          {formatTimeCompact(session.lastActivityAt)}
-        </span>
+        <span className="session-detail__time">{formatTimeCompact(session.lastActivityAt)}</span>
       </div>
 
-      {/* Terminal fills the viewport */}
       <div className={`session-detail__terminal-full${pr ? " session-detail__terminal-full--with-sheet" : ""}`}>
         {!showTerminal ? (
           <div className="session-detail-terminal-placeholder session-detail-height--full" />
         ) : terminalEnded ? (
           <div className="terminal-exited-placeholder session-detail-height--full">
-            <span className="terminal-exited-placeholder__text">
-              Terminal session has ended
-            </span>
+            <span className="terminal-exited-placeholder__text">Terminal session has ended</span>
           </div>
         ) : (
           <DirectTerminal
@@ -771,7 +360,6 @@ export function SessionDetail({
         )}
       </div>
 
-      {/* Bottom sheet with PR info */}
       {pr ? (
         <div className="session-detail__bottom-sheet">
           <div className="session-detail__sheet-handle" />
@@ -785,14 +373,10 @@ export function SessionDetail({
               PR #{pr.number}
             </a>
             <span className="session-detail__sheet-item">
-              <span
-                className={cn("session-detail__sheet-ci-dot", ciToneClass(pr))}
-              />
+              <span className={cn("session-detail__sheet-ci-dot", ciToneClass(pr))} />
               {getCiShortLabel(pr)}
             </span>
-            <span className="session-detail__sheet-item">
-              {getReviewShortLabel(pr) || "—"}
-            </span>
+            <span className="session-detail__sheet-item">{getReviewShortLabel(pr) || "—"}</span>
           </div>
         </div>
       ) : null}
@@ -807,331 +391,4 @@ export function SessionDetail({
       />
     </div>
   );
-}
-
-// ── Session detail PR card ────────────────────────────────────────────
-
-function SessionDetailPRCard({
-  pr,
-  sessionId,
-  metadata,
-  lifecyclePrReason,
-}: {
-  pr: DashboardPR;
-  sessionId: string;
-  metadata: Record<string, string>;
-  lifecyclePrReason?: string;
-}) {
-  const [sendingComments, setSendingComments] = useState<Set<string>>(new Set());
-  const [sentComments, setSentComments] = useState<Set<string>>(new Set());
-  const [errorComments, setErrorComments] = useState<Set<string>>(new Set());
-  const timersRef = useRef<Map<string, ReturnType<typeof setTimeout>>>(new Map());
-
-  useEffect(() => {
-    return () => {
-      timersRef.current.forEach((timer) => clearTimeout(timer));
-      timersRef.current.clear();
-    };
-  }, []);
-
-  const handleAskAgentToFix = async (comment: { url: string; path: string; body: string }) => {
-    setSentComments((prev) => {
-      const next = new Set(prev);
-      next.delete(comment.url);
-      return next;
-    });
-    setErrorComments((prev) => {
-      const next = new Set(prev);
-      next.delete(comment.url);
-      return next;
-    });
-    setSendingComments((prev) => new Set(prev).add(comment.url));
-
-    await askAgentToFix(
-      sessionId,
-      comment,
-      () => {
-        setSendingComments((prev) => {
-          const next = new Set(prev);
-          next.delete(comment.url);
-          return next;
-        });
-        setSentComments((prev) => new Set(prev).add(comment.url));
-        const existing = timersRef.current.get(comment.url);
-        if (existing) clearTimeout(existing);
-        const timer = setTimeout(() => {
-          setSentComments((prev) => {
-            const next = new Set(prev);
-            next.delete(comment.url);
-            return next;
-          });
-          timersRef.current.delete(comment.url);
-        }, 3000);
-        timersRef.current.set(comment.url, timer);
-      },
-      () => {
-        setSendingComments((prev) => {
-          const next = new Set(prev);
-          next.delete(comment.url);
-          return next;
-        });
-        setErrorComments((prev) => new Set(prev).add(comment.url));
-        const existing = timersRef.current.get(comment.url);
-        if (existing) clearTimeout(existing);
-        const timer = setTimeout(() => {
-          setErrorComments((prev) => {
-            const next = new Set(prev);
-            next.delete(comment.url);
-            return next;
-          });
-          timersRef.current.delete(comment.url);
-        }, 3000);
-        timersRef.current.set(comment.url, timer);
-      },
-    );
-  };
-
-  const allGreen = isPRMergeReady(pr);
-  const blockerIssues = buildBlockerChips(pr, metadata, lifecyclePrReason);
-  const fileCount = pr.changedFiles ?? 0;
-
-  return (
-    <div className={cn("session-detail-pr-card", allGreen && "session-detail-pr-card--green")}>
-      {/* Row 1: Title + diff stats */}
-      <div className="session-detail-pr-card__row">
-        <a
-          href={pr.url}
-          target="_blank"
-          rel="noopener noreferrer"
-          className="session-detail-pr-card__title-link"
-        >
-          PR #{pr.number}: {pr.title}
-        </a>
-        <span className="session-detail-pr-card__diff-stats">
-          <span className="session-detail-diff--add">+{pr.additions}</span>{" "}
-          <span className="session-detail-diff--del">-{pr.deletions}</span>
-        </span>
-        {fileCount > 0 && (
-          <span className="session-detail-pr-card__diff-label">
-            {fileCount} file{fileCount !== 1 ? "s" : ""}
-          </span>
-        )}
-        {pr.isDraft && (
-          <span className="session-detail-pr-card__diff-label">Draft</span>
-        )}
-        {pr.state === "merged" && (
-          <span className="session-detail-pr-card__diff-label">Merged</span>
-        )}
-      </div>
-
-      {/* Row 2: Blocker chips + CI chips inline */}
-      <div className="session-detail-pr-card__details">
-        {allGreen ? (
-          <div className="session-detail-merge-banner">
-            <svg
-              width="11"
-              height="11"
-              fill="none"
-              stroke="currentColor"
-              strokeWidth="2.5"
-              viewBox="0 0 24 24"
-            >
-              <path d="M20 6L9 17l-5-5" />
-            </svg>
-            Ready to merge
-          </div>
-        ) : (
-          blockerIssues.map((issue) => (
-            <span
-              key={issue.text}
-              className={cn(
-                "session-detail-blocker-chip",
-                issue.variant === "fail" && "session-detail-blocker-chip--fail",
-                issue.variant === "warn" && "session-detail-blocker-chip--warn",
-                issue.variant === "muted" && "session-detail-blocker-chip--muted",
-              )}
-            >
-              {issue.icon} {issue.text}
-              {issue.notified && (
-                <span className="session-detail-blocker-chip__note">· notified</span>
-              )}
-            </span>
-          ))
-        )}
-
-        {/* Separator between blockers and CI chips */}
-        {pr.ciChecks.length > 0 && (
-          <>
-            <div className="session-detail-pr-sep" />
-            {pr.ciChecks.map((check) => {
-              const chip = (
-                <span
-                  className={cn(
-                    "session-detail-ci-chip",
-                    check.status === "passed" && "session-detail-ci-chip--pass",
-                    check.status === "failed" && "session-detail-ci-chip--fail",
-                    check.status === "pending" && "session-detail-ci-chip--pending",
-                    check.status !== "passed" && check.status !== "failed" && check.status !== "pending" && "session-detail-ci-chip--queued",
-                  )}
-                >
-                  {check.status === "passed" ? "\u2713" : check.status === "failed" ? "\u2717" : check.status === "pending" ? "\u25CF" : "\u25CB"}{" "}
-                  {check.name}
-                </span>
-              );
-              return check.url ? (
-                <a
-                  key={check.name}
-                  href={check.url}
-                  target="_blank"
-                  rel="noopener noreferrer"
-                  className="hover:no-underline"
-                  onClick={(e) => e.stopPropagation()}
-                >
-                  {chip}
-                </a>
-              ) : (
-                <span key={check.name}>{chip}</span>
-              );
-            })}
-          </>
-        )}
-      </div>
-
-      {/* Row 3: Collapsible unresolved comments */}
-      {pr.unresolvedComments.length > 0 && (
-        <details className="session-detail-comments-strip">
-          <summary>
-            <div className="session-detail-comments-strip__toggle">
-              <svg
-                className="session-detail-comments-strip__chevron"
-                fill="none"
-                stroke="currentColor"
-                strokeWidth="2.5"
-                viewBox="0 0 24 24"
-              >
-                <path d="M9 5l7 7-7 7" />
-              </svg>
-              <span className="session-detail-comments-strip__label">Unresolved Comments</span>
-              <span className="session-detail-comments-strip__count">{pr.unresolvedThreads}</span>
-              <span className="session-detail-comments-strip__hint">click to expand</span>
-            </div>
-          </summary>
-          <div className="session-detail-comments-strip__body">
-            {pr.unresolvedComments.map((c, index) => {
-              const { title, description } = cleanBugbotComment(c.body);
-              return (
-                <details key={c.url} className="session-detail-comment" open={index === 0}>
-                  <summary>
-                    <div className="session-detail-comment__row">
-                      <svg
-                        className="session-detail-comment__chevron"
-                        fill="none"
-                        stroke="currentColor"
-                        strokeWidth="2"
-                        viewBox="0 0 24 24"
-                      >
-                        <path d="M9 5l7 7-7 7" />
-                      </svg>
-                      <span className="session-detail-comment__title">{title}</span>
-                      <span className="session-detail-comment__author">· {c.author}</span>
-                      <a
-                        href={c.url}
-                        target="_blank"
-                        rel="noopener noreferrer"
-                        onClick={(e) => e.stopPropagation()}
-                        className="session-detail-comment__view"
-                      >
-                        view &rarr;
-                      </a>
-                    </div>
-                  </summary>
-                  <div className="session-detail-comment__body">
-                    <div className="session-detail-comment__file">{c.path}</div>
-                    <p className="session-detail-comment__text">{description}</p>
-                    <button
-                      onClick={() => handleAskAgentToFix(c)}
-                      disabled={sendingComments.has(c.url)}
-                      className={cn(
-                        "session-detail-comment__fix-btn",
-                        sentComments.has(c.url) && "session-detail-comment__fix-btn--sent",
-                        errorComments.has(c.url) && "session-detail-comment__fix-btn--error",
-                      )}
-                    >
-                      {sendingComments.has(c.url)
-                        ? "Sending\u2026"
-                        : sentComments.has(c.url)
-                          ? "Sent \u2713"
-                          : errorComments.has(c.url)
-                            ? "Failed"
-                            : "Ask Agent to Fix"}
-                    </button>
-                  </div>
-                </details>
-              );
-            })}
-          </div>
-        </details>
-      )}
-    </div>
-  );
-}
-
-// ── Blocker chips helper (pre-merge blockers) ───────────────────────
-
-interface BlockerChip {
-  icon: string;
-  text: string;
-  variant: "fail" | "warn" | "muted";
-  notified?: boolean;
-}
-
-function buildBlockerChips(
-  pr: DashboardPR,
-  metadata: Record<string, string>,
-  lifecyclePrReason?: string,
-): BlockerChip[] {
-  const chips: BlockerChip[] = [];
-
-  const ciNotified = Boolean(metadata["lastCIFailureDispatchHash"]);
-  const conflictNotified = metadata["lastMergeConflictDispatched"] === "true";
-  const reviewNotified = Boolean(metadata["lastPendingReviewDispatchHash"]);
-  const lifecycleStatus = metadata["status"];
-
-  const ciIsFailing =
-    pr.ciStatus === CI_STATUS.FAILING ||
-    lifecyclePrReason === "ci_failing" ||
-    lifecycleStatus === "ci_failed";
-  const hasChangesRequested =
-    pr.reviewDecision === "changes_requested" ||
-    lifecyclePrReason === "changes_requested" ||
-    lifecycleStatus === "changes_requested";
-  const hasConflicts = pr.state !== "merged" && !pr.mergeability.noConflicts;
-
-  if (ciIsFailing) {
-    const failCount = pr.ciChecks.filter((c) => c.status === "failed").length;
-    chips.push({
-      icon: "\u2717",
-      variant: "fail",
-      text: failCount > 0 ? `${failCount} check${failCount !== 1 ? "s" : ""} failing` : "CI failing",
-      notified: ciNotified,
-    });
-  } else if (pr.ciStatus === CI_STATUS.PENDING) {
-    chips.push({ icon: "\u25CF", variant: "warn", text: "CI pending" });
-  }
-
-  if (hasChangesRequested) {
-    chips.push({ icon: "\u2717", variant: "fail", text: "Changes requested", notified: reviewNotified });
-  } else if (!pr.mergeability.approved) {
-    chips.push({ icon: "\u25CB", variant: "muted", text: "Awaiting reviewer" });
-  }
-
-  if (hasConflicts) {
-    chips.push({ icon: "\u2717", variant: "fail", text: "Merge conflicts", notified: conflictNotified });
-  }
-
-  if (pr.isDraft) {
-    chips.push({ icon: "\u25CB", variant: "muted", text: "Draft" });
-  }
-
-  return chips;
 }
