+++ packages/web/src/components/session-detail-agent-actions.ts
@@ -1,0 +1,23 @@
+import { cleanBugbotComment } from "./session-detail-utils";
+
+export async function askAgentToFix(
+  sessionId: string,
+  comment: { url: string; path: string; body: string },
+  onSuccess: () => void,
+  onError: () => void,
+) {
+  try {
+    const { title, description } = cleanBugbotComment(comment.body);
+    const message = `Please address this review comment:\n\nFile: ${comment.path}\nComment: ${title}\nDescription: ${description}\n\nComment URL: ${comment.url}\n\nAfter fixing, mark the comment as resolved at ${comment.url}`;
+    const response = await fetch(`/api/sessions/${encodeURIComponent(sessionId)}/message`, {
+      method: "POST",
+      headers: { "Content-Type": "application/json" },
+      body: JSON.stringify({ message }),
+    });
+    if (!response.ok) throw new Error(`HTTP ${response.status}`);
+    onSuccess();
+  } catch (error) {
+    console.error("Failed to send message to agent:", error);
+    onError();
+  }
+}
